app/bot.py
import os
import asyncio
from dotenv import load_dotenv
from aiogram import Bot, Dispatcher
from handlers import router as user_router


async def main():
    load_dotenv()

    bot = Bot(token=os.getenv("BOT_TOKEN"))
    dp = Dispatcher()

    dp.include_routers(user_router)
    await dp.start_polling(bot)


if __name__ == "__main__":
    try:
        asyncio.run(main())
    except KeyboardInterrupt:
        print("Stop Bot")
